Currency Converter: Privacy policy
This file represents the privacy policy of my Currency Converter app for Android which soon may be found on Google Play.

I created this app as a student for the simple use of converting units of currencies into one another as my first little Android Studio project. A list of the supported currencies with their momentary value in comparison to EUR is also provided. 

Additional features: 
- A click on one of those items leads to the Google Maps link which shows the capital city of the corresponding "home country" of respective currency
- Your phone notifies you when the used database for the supported currency values has been updated

I have not programmed this app to collect any personally identifiable information. Notifications by this app should be able to be turned off and the app can be uninstalled easily.

If you suspect any kind of harm caused by me or have questions about this app, you can contact me with the following e-mail adress: byteparkouring@t-online.de